fix: components refactor
