+++ packages/react-food-ya/src/components/Banner/index.js
@@ -1,5 +1,6 @@
 import { React, Component } from 'react';
-import Button from "./Button";
+import Button from "../Button";
+import './style.scss';
 
 class Banner extends Component {
     render() {
