+++ packages/react-food-ya/src/components/Banner/style.scss
@@ -1,0 +1,1 @@
+@import "../../styles/variables";
