+++ packages/react-food-ya/src/components/Button/style.scss
@@ -1,0 +1,11 @@
+@import "../../styles/variables";
+
+.default-button {
+    width: 150px;
+    height: 50px;
+    background: $color_green;
+    padding: 10px 25px;
+    border-radius: 10px;
+    color: white;
+    text-decoration: none;
+}
