+++ packages/react-food-ya/src/components/Footer/index.js
