+++ packages/react-food-ya/src/components/GridCard/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from './Button';
+import Button from '../Button/';
 
 class GridCard extends React.Component {
     render() {
