+++ packages/react-food-ya/src/styles/index.scss
@@ -9,16 +9,6 @@ body {
     background: $color_cream;
 }
 
-.default-button {
-    width: 150px;
-    height: 50px;
-    background: $color_green;
-    padding: 10px 25px;
-    border-radius: 10px;
-    color: white;
-    text-decoration: none;
-}
-
 /* Navbar */
 
 /* Banner */
